Create a differ aspect
I was discussing Discussing the tool the golden parser with a friend of mine ...not very often do i get to actually "talk" to another person about this stuff and the way he was describing how it worked ?I thought that sounded crazy ? that not parsing that go fish and match? guess matching at that ? parsing is taking know data like a login as an example and using that know login form one source to the next(IF THE RELATIONSHIP OF THE  LOGIN IS KNOWN TO BE TRUE ) you can use one log on any site that has matching log info ... OOH YES IT DOES ANY SITE  IVE tried to give it in a bounty no one will take it? apparently I'm not credible enough  BUT STILL THAT BEING SAID IM NOT RELEASE THE EXACT DETAILS AS IS WITH ALL MY  YOUR BOUNTY IS MY DOLLA! TILL NEXT TTIME